devide jobs and choose the correct general status
