+++ app/next-client-app/components/jobs/JobUtils.tsx
@@ -17,3 +17,24 @@ export const FindGeneralStatus = (jobsData: Job[]) => {
   }
   return generalStatus;
 };
+
+export const DivideJobs = (jobsData: Job[]) => {
+  let jobGroups: Job[][] = [];
+  // Divide the jobs of each table to group of three (each group demonstrates each run)
+  if (jobsData.length > 0) {
+    let jobs: Job[] = [];
+    jobsData.forEach((job) => {
+      jobs.push(job);
+      if (jobs.length === 3) {
+        // Sort jobs based on the "created_at" field
+        jobs.sort(
+          (a, b) =>
+            new Date(a.created_at).getTime() - new Date(b.created_at).getTime()
+        );
+        jobGroups.push(jobs);
+        jobs = [];
+      }
+    });
+  }
+  return jobGroups;
+};
